refactor: rename transact_previous to replay, move some traits
