+++ bins/revme/src/cmd/evmrunner.rs
@@ -97,7 +97,7 @@ impl Cmd {
             let bench_options = microbench::Options::default().time(Duration::from_secs(3));
 
             microbench::bench(&bench_options, "Run bytecode", || {
-                let _ = evm.transact_previous().unwrap();
+                let _ = evm.replay().unwrap();
             });
 
             return Ok(());
@@ -106,7 +106,7 @@ impl Cmd {
         let out = if self.trace {
             evm.inspect_previous().map_err(|_| Errors::EVMError)?
         } else {
-            let out = evm.transact_previous().map_err(|_| Errors::EVMError)?;
+            let out = evm.replay().map_err(|_| Errors::EVMError)?;
             println!("Result: {:#?}", out.result);
             out
         };
