+++ crates/context/src/setters.rs
@@ -2,12 +2,18 @@ use crate::Context;
 use auto_impl::auto_impl;
 use context_interface::{Block, Cfg, Database, Journal, Transaction};
 
+/// Setters for the context.
 #[auto_impl(&mut, Box)]
 pub trait ContextSetters {
+    /// Transaction type.
     type Tx: Transaction;
+    /// Block type.
     type Block: Block;
 
+    /// Set the transaction.
     fn set_tx(&mut self, tx: Self::Tx);
+
+    /// Set the block.
     fn set_block(&mut self, block: Self::Block);
 }
